refactor(entities): update imports
Add support for `@/x` kind of import for explicit cross export/import
Same level imports use relative paths
