+++ src/entities/author/@x/selectors.ts
@@ -1,0 +1,1 @@
+export { selectAuthorByID } from "../model/slice";
