+++ src/entities/tweet/@x/types.ts
@@ -1,0 +1,1 @@
+export type { Tweet as TweetType } from "../model/types";
