+++ src/entities/user/@x/selectors.ts
@@ -1,0 +1,1 @@
+export { selectCurrentUser, selectUserDataByID } from "../model/slice";
