+++ src/entities/user/@x/types.ts
@@ -1,0 +1,1 @@
+export type { UserData, UserState } from "../model/types";
