+++ src/entities/user/api/updateUserData.ts
@@ -1,5 +1,5 @@
-import type { UserDataUpdate } from "@/entities/user";
 import { isDataExists, updateData } from "@/shared/lib/firebase";
+import { UserDataUpdate } from "..";
 
 export const updateUserData = async (userId: string, data: UserDataUpdate) => {
   const dbPath = "users/" + userId;
